move app.js file to components folder
