+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\Avijit\\React\\math-magicians-calculator\\src\\index.js":"1","C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\Button.js":"2","C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\ButtonPanel.js":"3","C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\Display.js":"4","C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\App.js":"5"},{"size":231,"mtime":1611941899552,"results":"6","hashOfConfig":"7"},{"size":309,"mtime":1611939673474,"results":"8","hashOfConfig":"7"},{"size":876,"mtime":1611941178590,"results":"9","hashOfConfig":"7"},{"size":325,"mtime":1611940576769,"results":"10","hashOfConfig":"7"},{"size":419,"mtime":1611941919273,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1idw9eu",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Avijit\\React\\math-magicians-calculator\\src\\index.js",[],"C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\Button.js",[],"C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\ButtonPanel.js",[],"C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\Display.js",[],"C:\\Avijit\\React\\math-magicians-calculator\\src\\components\\App.js",[]]
