+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import Display from './components/Display';
-import ButtonPanel from './components/ButtonPanel';
+import Display from './Display';
+import ButtonPanel from './ButtonPanel';
 
 class App extends Component {
   constructor(props) {
